testing/web-platform/mozilla/tests/indic-detection/htchanakya.html

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
„ії№ їлг вхЏU„і«№ „рі вкЎь вр њхЌж їэг гс– ІвXWр ћ№ XWи •ж÷ж ЏUзQ„ гс, зЅв яЅг вр Івр "№ж№ їэг" XWр ўж„ вр ÷и Ѕжўж Ѕжћж гс– вхЏU„і«№ XWр їэг ќф ћЏUг XWр гфћр гсі - "SЌ№иЎ їэг" зЅўXWж ћ№ •ж÷жвиЎ гфћж гс •хЏU "їсвиЎ їэг" Ѕф •зѕXWћЏU їсв вр зўз„ьћ гсі– ¬лfяи XWи ћЏUг, „ії№ ÷и ∞XW SЌ№иЎ ѕЏUжћ№ яж№ж їэг гс– ІвXWж яжћжяЏUJж зяЏU№ гс– ІвXWи вћг ќрєўр ¬ЏU њіќэ„ж XWр їћь •хЏU ¬лfяи XWр —яж№ж„йзєЎфі, ≤жжз≈Ўфі, ЏUрзїSћжў •хЏU VLWяиЎ ’YWиь№и њфз≈Ўфі XWи Ўжќ зќ№жћи гс– вхЏU„і«№ XWж в’вр •зѕXW ™W°њж ¬яьћ, •ф№≥¬в „фiв „ії№ ¬ЏU ги зSЌћ гс– вжЌ ги зяажж№ћ„ XWсiЎў яс№рв „сЏUиўрзЏUв ÷и Ўгиі ¬ЏU зSЌћ гс– •¬ўи ÷хїфз№XW зяажрбћж•фі XWр •№жяж, „ії№ XWж ≤жкJжьў XWж№ •хЏU „хв„и њXWэ ¬лfяи XWр в„жў гсі– Ів їэг ¬ЏU Ѕияў гфўр XWи ві÷жяўж XWф г„ражж вр ¬зЏUXWзЋ¬ћ зXWЎж їЎж гс–

v~{z „рі „рзЏUўЏU y XWр mжЏUж XWи ¬г№и „ії№ ©«жў вр ¬г№р ћXW Ўг „жўж Ѕжћж Ќж зXW їэг XWи вћг ¬ЏU ћЏU№ •яSЌж „рі Ѕ№ гф вXWћж гс– Ўг гЋXWр •хЏU їгЏUр ЏUії XWр ѕ¶’фі XWи •жязћьXW вкњўж•фі ¬ЏU •жѕжзЏUћ Ќж зяажрб ћхЏU ¬ЏU, VLWяиЎ •€жжіажфі, Ѕф №і’р гфўр ¬ЏU в„йќэ •хЏU „гжmи¬фі XWи ћЏUг зќєћр гсі, XWж№р striations XWи √ЎжGЎж XWйјU ¬эр€жXWфі mжЏUж ¬жўи XWи звіњжІь ўгЏUфі XWр MW¬ „рі XWи їЎи гс– Іўч виѕи ЏUрєж•фі XWи „хЅкќїи ’жќ „рі звh ўгиі гф ¬жЎи •хЏU Ўр „жўж їЎж зXW Ўр ЏUрєжЎрі „жµж ¬эXWжажиЎ ÷э„ XWр •№жяж XWйјU •хЏU ўгиі гсі– зYWЏU ÷и, вхЏU „і«№ XWр в÷и їэгфі „рі ¬лfяи XWр •№жяж, „ії№ XWф Ѕияў зяSћжЏU XWж „гPя¬кJжь зяXWЋ¬ „жўж Ѕжћж гс–
<script>
test(function() {
  assert_equals(document.characterSet, "windows-1252");
},"Should fall back to windows-1252");
</script>
The text content above is a converted extract from the start of the Wikipedia article <a href="https://hi.wikipedia.org/w/index.php?title=%E0%A4%AE%E0%A4%82%E0%A4%97%E0%A4%B2_%E0%A4%97%E0%A5%8D%E0%A4%B0%E0%A4%B9&oldid=5105576">&#2350;&#2306;&#2327;&#2354; &#2327;&#2381;&#2352;&#2361;</a> and
is licensed under the <a href="https://creativecommons.org/licenses/by-sa/3.0/legalcode">Creative Commons Attribution-ShareAlike 3.0 Unported</a> license.
